+++ src/shared/constants.js
@@ -8,7 +8,8 @@ import { __ } from 'i18n';
 export const VIEW_PLUGINS = [
     'converse-adhoc-views',
     'converse-bookmark-views',
-    'converse-chatboxviews',
+    'converse-app-chat',
+    'converse-app-social',
     'converse-chatview',
     'converse-controlbox',
     'converse-disco-views',