+++ src/types/shared/texture/utils.d.ts
@@ -90,13 +90,13 @@ declare const _default: {
     isString(s: any): boolean;
     getDefaultStorageType(): import("headless/types/utils/types.js").StorageType;
     isPersistentStorageAvailable(): boolean;
-    createStore(id: string, type: import("headless/types/utils/types.js").StorageType): import("@converse/skeletor").PersistentStorage;
+    createStore(id: string, type?: import("headless/types/utils/types.js").StorageType): import("@converse/skeletor").PersistentStorage;
     initStorage(model: import("headless/types/utils/types.js").StorageModel, id: string, type?: import("headless/types/utils/types.js").StorageType): void;
     isErrorStanza(stanza: Element): boolean;
     isForbiddenError(stanza: Element): boolean;
     isServiceUnavailableError(stanza: Element): boolean;
     getAttributes(stanza: Element): object;
-    toStanza: typeof import("@converse/headless").Stanza.toElement;
+    toStanza: typeof import("strophe.js").Stanza.toElement;
     isUniView(): boolean;
     isTestEnv(): boolean;
     getUnloadEvent(): "pagehide" | "beforeunload" | "unload";
@@ -122,17 +122,11 @@ declare const _default: {
     isSameBareJID(jid1: string, jid2: string): boolean;
     isSameDomain(jid1: string, jid2: string): boolean;
     getJIDFromURI(jid: string): string;
+    getNodeFromURI(uri?: string): string | undefined;
+    getItemFromURI(uri?: string): string | undefined;
     isOwnJID(jid: string, include_resource?: boolean): boolean;
     maybeAppendDomain(jid: string): string;
-    initPlugins(_converse: 
-    /**
-     * Given a specific index "i" of "text", return the directive it matches or null otherwise.
-     * @param {import('./texture').Texture} text - The text in which  the directive appears
-     * @param {Number} i - The directive index
-     * @param {Boolean} opening - Whether we're looking for an opening or closing directive
-     * @returns {string|null}
-     */
-    ConversePrivateGlobal): void;
+    initPlugins(_converse: ConversePrivateGlobal): void;
     initClientConfig(_converse: ConversePrivateGlobal): Promise<void>;
     initSessionStorage(_converse: ConversePrivateGlobal): Promise<void>;
     initPersistentStorage(_converse: ConversePrivateGlobal, store_name: string, key?: string): void;
@@ -145,7 +139,7 @@ declare const _default: {
     safeSave(model: import("@converse/skeletor").Model, attributes: any, options: any): void;
     isElement(el: unknown): boolean;
     isEqualNode(actual: Element, expected: Element): boolean;
-    isTagEqual(stanza: Element | typeof import("@converse/headless").Builder, name: string): boolean;
+    isTagEqual(stanza: Element | typeof import("strophe.js").Builder, name: string): boolean;
     stringToElement(s: string): Element;
     queryChildren(el: HTMLElement, selector: string): ChildNode[];
     siblingIndex(el: Element): number;